+++ packages/desktop/src/renderer/components/MathMessage/customMathJax.ts
@@ -1,34 +1,3 @@
-// // import mathjax from 'mathjax-full/js/mathjax'
-// var mathjax = require('mathjax-full/js/mathjax');
-// var tex$1 = require('mathjax-full/js/input/tex');
-// var mathml$1 = require('mathjax-full/js/input/mathml');
-// var svg$1 = require('mathjax-full/js/output/svg');
-// var browserAdaptor = require('mathjax-full/js/adaptors/browserAdaptor');
-// var html = require('mathjax-full/js/handlers/html');
-// var MathItem = require('mathjax-full/js/core/MathItem');
-
-// const onError = (e) => {console.log('ERROR', e)}
-// // create and register adaptor bound to the real DOM
-// var adaptor = browserAdaptor.browserAdaptor();
-// html.RegisterHTMLHandler(adaptor);
-
-// var tex = new tex$1.TeX({ packages: ["base", "ams"] });
-// // var mathml = new mathml$1.MathML({});
-// var svg = new svg$1.SVG({ fontCache: "none" });
-// var markErrors = [MathItem.STATE.TYPESET + 1, null, onError];
-
-// var tex_html = mathjax.mathjax.document("", {
-//     InputJax: tex,
-//     OutputJax: svg,
-//     renderActions: {
-//         markErrors: markErrors,
-//     },
-// });
-
-
-
-
-
 'use strict';
 
 Object.defineProperty(exports, '__esModule', { value: true });
@@ -63,7 +32,6 @@ function _interopNamespace(e) {
 export function convertToSvg(src, display, settings) {
     const math = src.trim();
     const converted = tex_html.convert(math, { display: display, ...settings })
-    // return converted
     tex_html.updateDocument()
     return adaptor.outerHTML(converted)
 }
@@ -140,48 +108,6 @@ export function convertPromise(srcSpec, node, display, settings) {
         // do stuff with dom
         html.updateDocument();
         updateCSS("MATHJAX-SVG-STYLESHEET", svg.cssStyles.cssText);
-        var err = adaptor.getAttribute(dom, "data-mjx-error");
-        if (err) {
-            throw err;
-        }
-        return adaptor.outerHTML(dom);
-    })
-        .catch(function (err) {
-        if (!(err instanceof CancelationException)) {
-            throw err;
-        }
-        else {
-            console.log("cancelled render!");
-        }
-    });
-    return { promise: res.then(function (v) { return (v ? v : ""); }), cancel: cancel };
-}
-
-
-export function customConvertPromise(srcSpec, node, display, settings) {
-    var src = srcSpec.src//, lang = srcSpec.lang;
-    // if (!node)
-    //     throw new Error();
-    // var html = tex_html;
-    // if (lang == "MathML")
-    //     html = mathml_html;
-    var math = src.trim();
-    // const metrics = svg.getMetricsFor(node, display);
-    var canceled = false;
-    var cancel = function () { return (canceled = true); };
-    var res = mathjax.mathjax
-        .handleRetriesFor(function () {
-        if (canceled) {
-            throw new CancelationException();
-        }
-        // var dom = html.convert(math, __assign({ display: display }, settings));
-        var dom = tex_html.convert(math, { display: display, ...settings });
-        return dom;
-    })
-        .then(function (dom) {
-        // do stuff with dom
-        tex_html.updateDocument();
-        // updateCSS("MATHJAX-SVG-STYLESHEET", svg.cssStyles.cssText);
         var err = adaptor.getAttribute(dom, "data-mjx-error");
         if (err) {
             throw err;
